+++ _build/build.transport.php
@@ -15,8 +15,8 @@
  * details.
  *
  * You should have received a copy of the GNU General Public License along with
- * Rowboat; if not, write to the Free Software Foundation, Inc., 59 Temple
- * Place, Suite 330, Boston, MA 02111-1307 USA
+ * minifyRegistered; if not, write to the Free Software Foundation, Inc., 
+ * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
  * @package minifyregistered
  * @subpackage build
@@ -32,7 +32,7 @@ set_time_limit(0);
 /* define package */
 define('PKG_NAME', 'minifyRegistered');
 define('PKG_NAME_LOWER', strtolower(PKG_NAME));
-define('PKG_VERSION', '0.2.1');
+define('PKG_VERSION', '0.3.0');
 define('PKG_RELEASE', 'pl');
 
 /* define sources */
@@ -40,17 +40,19 @@ $root = dirname(dirname(__FILE__)) . '/';
 $sources = array(
 	'root' => $root,
 	'build' => $root . '_build/',
-	'data' => $root . '_build/data/',
 	'resolvers' => $root . '_build/resolvers/',
-	'properties' => $root . '_build/properties/',
-	'chunks' => $root . 'core/components/' . PKG_NAME_LOWER . '/elements/chunks/',
-	'snippets' => $root . 'core/components/' . PKG_NAME_LOWER . '/elements/snippets/',
+	'data' => $root . '_build/data/',
+    'events' => $root . '_build/data/events/',
+    'permissions' => $root . '_build/data/permissions/',
+    'properties' => $root . '_build/data/properties/',
+	'source_core' => $root . 'core/components/' . PKG_NAME_LOWER,
+	'source_assets' => $root . 'assets/components/' . PKG_NAME_LOWER,
+	'source_min' => $root . 'assets/min',
 	'plugins' => $root . 'core/components/' . PKG_NAME_LOWER . '/elements/plugins/',
+	'snippets' => $root . 'core/components/' . PKG_NAME_LOWER . '/elements/snippets/',
+	'chunks' => $root . 'core/components/' . PKG_NAME_LOWER . '/elements/chunks/',
 	'lexicon' => $root . 'core/components/' . PKG_NAME_LOWER . '/lexicon/',
 	'docs' => $root . 'core/components/' . PKG_NAME_LOWER . '/docs/',
-	'pages' => $root . 'core/components/' . PKG_NAME_LOWER . '/elements/pages/',
-	'source_assets' => $root . 'assets/components/' . PKG_NAME_LOWER,
-	'source_core' => $root . 'core/components/' . PKG_NAME_LOWER,
 );
 unset($root);
 
@@ -70,28 +72,29 @@ $builder = new modPackageBuilder($modx);
 $builder->createPackage(PKG_NAME_LOWER, PKG_VERSION, PKG_RELEASE);
 $builder->registerNamespace(PKG_NAME_LOWER, false, true, '{core_path}components/' . PKG_NAME_LOWER . '/');
 
-/* add plugin */
-$modx->log(modX::LOG_LEVEL_INFO, 'Packaging in plugin...');
-$plugin = $modx->newObject('modPlugin');
-$plugin->fromArray(array(
-	'id' => 1,
-	'name' => 'minifyRegistered',
-	'description' => 'Collect the registered javascript and css files/chunks and minify them.',
-	'plugincode' => getSnippetContent($sources['plugins'] . 'plugin.minifyregistered.php'),
-		), '', true, true);
-$events = array();
-$events['OnWebPagePrerender'] = $modx->newObject('modPluginEvent');
-$events['OnWebPagePrerender']->fromArray(array(
-	'event' => 'OnWebPagePrerender',
-	'priority' => 0,
-	'propertyset' => 0,
-		), '', true, true);
-$plugin->addMany($events);
-unset($events);
-$properties = include $sources['properties'] . 'properties.minifyregistered.php';
-$plugin->setProperties($properties);
-unset($properties);
+/* load system settings */
+$settings = include_once $sources['data'] . 'transport.settings.php';
+$attributes = array(
+	xPDOTransport::UNIQUE_KEY => 'key',
+	xPDOTransport::PRESERVE_KEYS => true,
+	xPDOTransport::UPDATE_OBJECT => false,
+);
+if (!is_array($settings)) {
+	$modx->log(modX::LOG_LEVEL_FATAL, 'Adding settings failed.');
+}
+foreach ($settings as $setting) {
+	$vehicle = $builder->createVehicle($setting, $attributes);
+	$builder->putVehicle($vehicle);
+}
+$modx->log(modX::LOG_LEVEL_INFO, 'Packaged in ' . count($settings) . ' system settings.');
+flush();
+unset($settings, $setting, $attributes);
 
+/* add plugins */
+$plugins = include $sources['data'] . 'transport.plugins.php';
+if (!is_array($plugins)) {
+	$modx->log(modX::LOG_LEVEL_FATAL, 'Adding plugins failed.');
+}
 $attributes = array(
 	xPDOTransport::UNIQUE_KEY => 'name',
 	xPDOTransport::PRESERVE_KEYS => false,
@@ -105,13 +108,46 @@ $attributes = array(
 		),
 	),
 );
-$vehicle = $builder->createVehicle($plugin, $attributes);
+foreach ($plugins as $plugin) {
+	$vehicle = $builder->createVehicle($plugin, $attributes);
+	$builder->putVehicle($vehicle);
+}
+$modx->log(modX::LOG_LEVEL_INFO, 'Packaged in ' . count($plugins) . ' plugins.');
+flush();
+unset($plugins, $plugin, $attributes);
 
-$modx->log(modX::LOG_LEVEL_INFO, 'Adding file resolvers to plugin...');
+/* create category */
+$category = $modx->newObject('modCategory');
+$category->set('id', 1);
+$category->set('category', PKG_NAME);
+$modx->log(modX::LOG_LEVEL_INFO, 'Packaged in category.');
+flush();
+
+/* create category vehicle */
+$attr = array(
+	xPDOTransport::UNIQUE_KEY => 'category',
+	xPDOTransport::PRESERVE_KEYS => false,
+	xPDOTransport::UPDATE_OBJECT => true,
+	xPDOTransport::RELATED_OBJECTS => true,
+	xPDOTransport::RELATED_OBJECT_ATTRIBUTES => array(
+		'Snippets' => array(
+			xPDOTransport::PRESERVE_KEYS => false,
+			xPDOTransport::UPDATE_OBJECT => true,
+			xPDOTransport::UNIQUE_KEY => 'name',
+		),
+	)
+);
+$vehicle = $builder->createVehicle($category, $attr);
 $vehicle->resolve('file', array(
 	'source' => $sources['source_core'],
-	'target' => "return MODX_CORE_PATH . 'components/';",
+	'target' => "return MODX_CORE_PATH . 'components/';"
 ));
+$vehicle->resolve('file', array(
+	'source' => $sources['source_min'],
+	'target' => "return MODX_ASSETS_PATH;"
+));
+$modx->log(modX::LOG_LEVEL_INFO, 'Packaged in folders.');
+flush();
 $builder->putVehicle($vehicle);
 
 /* now pack in the license file, readme and changelog */