+++ _build/includes/functions.php
@@ -15,8 +15,8 @@
  * details.
  *
  * You should have received a copy of the GNU General Public License along with
- * Rowboat; if not, write to the Free Software Foundation, Inc., 59 Temple
- * Place, Suite 330, Boston, MA 02111-1307 USA
+ * minifyRegistered; if not, write to the Free Software Foundation, Inc., 
+ * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
  * @package minifyregistered
  * @subpackage build